Add initial swedish colemak keymap for corne
This is my personal keymap, and will likely be modified further. It is
very much inspired by colemak, but adapted for the corne keyboard. To
be able to write the swedish characters ÄÅÖäåö the keymap is written for
software using the swedish keyboard layout. But to preserve the
US/colemak shifted keys, we have overridden shift keys for the first
layer.

It still has some issues and uncertainties, such as the thumb keys,
space, backspace, escape.